Fixed buffer type error on linux

+++ run.py
@@ -23,7 +23,7 @@ for i in xrange(numPoints):
     
 # Get polygon connectivity
 numPolys     = stl.GetNumberOfCells()
-connectivity = np.zeros((numPolys,3),dtype=int)
+connectivity = np.zeros((numPolys,3),dtype=np.int32)
 for i in xrange(numPolys):
     tri = stl.GetCell(i)
     ids = tri.GetPointIds()
